:recycle: Refactor: CommonSnackBar :core:ui 모듈로 이동
